telegram-daily-bot/requirements.txt
python-telegram-bot[job-queue]==22.0
python-dotenv==1.0.1
